+++ app/routes/product-details.$productSlug/route.module.scss
